Add support to portuguese keywords
Just hacked the lexer to translate the keywords
for their counterparts in english and return
the right tokens to the parser. It supports the
original spec because it uses the previous compiled
javac in order to compile its own standard library.

Also, had to modify some code, because some parts of the
library used some words that are actually portuguese translations
for some keywords, like `classe` and `se`.